src/mymodel-cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mymodel"
version = "0.1.0"
description = "Create your personal AI model from multiple LLMs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
keywords = ["mymodel", "llm", "routing", "ai", "proxy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.7",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "huggingface_hub[cli]>=0.20.0",
    "textual>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "black>=22.0.0",
]

[project.scripts]
mymodel = "mymodel.cli.main:main"

[tool.setuptools]
packages = { find = { where = ["."] } }

[tool.setuptools.package-data]
"mymodel.cli.templates" = ["*.yaml", "*.json", "*.ini"]
"mymodel.cli.styles" = ["*.tcss"]

[tool.pytest.ini_options]
addopts = ["-v"]
testpaths = ["tests"]
